+++ src/com/rian/difficultycalculator/hitobject/Spinner.java
@@ -14,7 +14,7 @@ public class Spinner extends HitObjectWithDuration {
      * @param timingControlPoint     The timing control point this spinner is under effect on.
      * @param difficultyControlPoint The difficulty control point this spinner is under effect on.
      */
-    public Spinner(int startTime, int endTime, TimingControlPoint timingControlPoint, DifficultyControlPoint difficultyControlPoint) {
+    public Spinner(double startTime, double endTime, TimingControlPoint timingControlPoint, DifficultyControlPoint difficultyControlPoint) {
         super(startTime, endTime, new Vector2(256, 192), timingControlPoint, difficultyControlPoint);
     }
 }
